removed repeated code
