+++ src/components/ShowMobileView.js
@@ -1,0 +1,13 @@
+import useAppStyle from "../appStyle";
+import { NavigationMob } from "./NavigationMob";
+
+export const ShowMobileView = (props) => {
+  const classes = useAppStyle();
+
+  return (
+    <div className={classes.pageContainer}>
+      <NavigationMob />
+      {props.element}
+    </div>
+  );
+};
